requirements.txt
# PTF Tahmin SaaS Platform - Requirements
# =========================================

# Django Core
Django>=4.2,<5.0
gunicorn>=21.0.0
whitenoise>=6.6.0

# Database
psycopg2-binary>=2.9.0

# Authentication
django-allauth>=0.58.0

# REST API
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0

# Celery (Background Tasks)
celery>=5.3.0
django-celery-beat>=2.5.0
redis>=5.0.0

# Forms
django-crispy-forms>=2.1
crispy-tailwind>=0.5.0

# ML & Data
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0

# EPİAŞ API
eptr2>=1.3.0

# Emtia Verileri
yfinance>=0.2.0
requests>=2.28.0

# Payment (iyzico)
# iyzipay>=2.0.0  # Production için uncomment

# Utils
python-dotenv>=1.0.0
PyYAML>=6.0

# Development
django-debug-toolbar>=4.2.0

# Production
sentry-sdk>=1.34.0
